server/requirements.txt
# ---- Core API ----
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.12.5
python-multipart==0.0.9
# bcrypt==4.3.0

# ---- Environment ----
python-dotenv==1.0.1

# ---- Langchain ----
langchain==1.2.0
langchain-community==0.4.1
langchain-text-splitters==1.1.0


# ---- LLM Providers ----
langchain-groq==1.1.1
langchain-openai==1.1.6

# ---- OpenAI ----
openai==2.14.0

# ---- Embeddings & Vector DB ----
faiss-cpu==1.13.0
tiktoken==0.7.0

# ---- Document Loaders ----
pypdf>=6.6.0
unstructured>=0.18.27
python-docx==1.1.2

# ---- Async / Performance ----
httpx==0.27.0

# ---- Optional (Recommended) ----
tenacity==8.2.3
loguru==0.7.2

# ---- Database ----
sqlalchemy==2.0.37
psycopg2-binary==2.9.10

# ---- Authentication ----
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose[cryptography]
PyJWT==2.8.0
alembic

# ---- Rate Limiting ----
slowapi==0.1.9
limits==5.6.0